fix: move EMealTypes to utils.ts to fix a circular dependency
